# Fix num_byte calculations to use nhead_k for K & V size