@cameronshinn 2025-08-09
Submitted a PR upstream here https://github.com/ROCm/composable_kernel/pull/2653